arch/aes: Import Intel assembler code for AES GCM.
The Haswell-optimized variant of Intel's AES GCM is imported. This is
optimized crypto code that is particularly suited for fast IPsec.

Intel's software distribution also includes other variants for older
CPUs but those are not included in this commit.

This code is based on the YASM assembler. This is not an ideal match
for Snabb Switch and it may be a worthwhile exercise to port this code
to DynASM to keep our build dependencies simple. One thing at a time...