connectx4: Reorganize SHM objects
There is now a single SHM frame containing all state for a send+receive
queue pair. (Previously there was one frame for send and one for
receive.)

The IO app also now polls for the SHM objects to become available and so
it does not need to be started before the ConnectX4 controller.