basic_apps: Add 'name' fields to disambiguate
The engine chooses a profiling "zone" based on a heuristic name for
the app class. This heuristic works poorly for the basic apps -
falling back to the module name that is the same for all of them - so
we add an explicit 'name' field to the app class to help out.