connectx4: Eliminate heap allocation
Couple of functions allocated pointer values at the start of the
function and then passed these values into the loop. LuaJIT is
compiling the function head and the loop into separate traces and so
when the function head trace finishes its snapshot forced the pointers
to be boxed and stored on the stack.

Fixed by making sure no pointers are bound to local variables at the
point where the loop starts.

(I suspect LuaJIT could handle this automatically if the loop were
either part of the function head trace or were a side-trace. I am
curious to look more closely at why that doesn't happen and whether
there is an easy way to make it so. One thing at at time though :-)